# api/swagger.yaml
openapi: 3.0.3
info:
  title: Athlete Hub API
  description: |
    REST API for the Athlete Hub sports events platform.

    ## Authentication
    - **Protected routes** require `Authorization: Bearer <session_token>`
    - **Public routes** require no authentication

    ## Base URL
    All endpoints are prefixed with `/api`.
  version: "1.0.0"
  contact:
    name: Athlete Hub Dev
servers:
  - url: /api
    description: Same-origin (development & production)
...
paths:
...
  /events/{slug}:
    get:
      tags: [Marketplace]
      summary: Get event detail with categories, registration fields, course, and media
      description: |
        Optional Bearer athlete token. When authenticated, includes `myRegistration`
        if the athlete has a pending or confirmed registration for this event.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event detail (may include myRegistration for signed-in athletes)
          content:
            application/json:
              schema:
                type: object
                properties:
                  feePresentation:
                    $ref: "#/components/schemas/FeePresentation"
                  payments_available:
                    type: boolean
                    description: False when published paid categories exist but organizer Connect payout is not ready
                  has_paid_categories:
                    type: boolean
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/EventCategoryPricing"
        "404":
          description: Event not found
...
  /organizer/events/{eventId}:
    get:
      tags: [Organizer]
      summary: Event detail with categories
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event detail (includes bib_mode)
    patch:
      tags: [Organizer]
      summary: Update organizer event
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffEventUpsertRequest"
      responses:
        "200":
          description: Updated event
        "400":
          description: Validation error (includes subdomain_immutable when subdomain changes)
    delete:
      tags: [Organizer]
      summary: Soft-delete event (cancels open registrations; prefer deactivate-listing)
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Soft-deleted; returns cancelledRegistrations count
        "403":
          description: Only owner/organizer
...
  /admin/events/{eventId}:
    get:
      tags: [Admin]
      summary: Event detail with categories
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event detail
    patch:
      tags: [Admin]
      summary: Update any event
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated event
    delete:
      tags: [Admin]
      summary: Soft-delete any event (cancels open registrations)
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Soft-deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EventBibMode:
      type: string
      enum: [folio, separate]
      description: >
        folio = registration folio is also the race bib (dorsal); separate = folio and bib independent.
        New events default to folio; existing events migrate to separate.
...
    StaffEventUpsertRequest:
      type: object
      description: >
        On create, title / subdomain / sport_type_id / start_date are required
        (enforced by API). On PATCH, subdomain is immutable — a different value
        returns 400 with code subdomain_immutable.
      properties:
        title:
          type: string
        sport_type_id:
          type: integer
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        slug:
          type: string
        subdomain:
          type: string
          description: >
            Required on create. Globally unique vanity host label for
            `{subdomain}.atleita.com`. Immutable after create.
          minLength: 3
          maxLength: 63
        short_description:
          type: string
        description:
          type: string
        visibility:
          type: string
          enum: [public, private, unlisted]
        featured:
          type: boolean
        registration_opens_at:
          type: string
          format: date-time
        registration_closes_at:
          type: string
          format: date-time
        location_name:
          type: string
        location_city:
          type: string
        location_state:
          type: string
        max_registrations:
          type: integer
        max_registrations_per_order:
          type: integer
        bib_mode:
          $ref: "#/components/schemas/EventBibMode"
        auto_deactivate_after_event:
          type: boolean
          description: When true (default), unlist from marketplace after the event day ends
        fee_presentation:
          $ref: "#/components/schemas/FeePresentation"
        msi_enabled:
          type: boolean
          description: Offer Stripe MSI (3/6/9) on checkout when payouts are ready
        manual_sales_enabled:
          type: boolean
          description: >
            When true, staff may record cash manual sales (provider=manual) for this event.
            Default false.
    FeePresentation:
      type: string
      enum: [pass_through, absorb_all]
...
    EventCategoryPricing:
      type: object
      properties:
        price_cents:
          type: integer
        total_cents:
          type: integer
          description: Athlete-facing checkout total for this category
        display_iva_cents:
          type: integer
        organizer_fiscal_net_cents:
          type: integer
...